Hack away at the basic style
The forked theme ended up using a massive css library which was all a
bit behond closed doors. I have attempted to rip that out and replace it
with hand written code.

It might not be as all covering as the framework but I will know what is
going on.